migrate export_caffe2_op_to_c10.h macros to the new dispatcher registration API, update code_analyzer regex
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48308

The original regex that I added didn't correctly match namespaces that started with an underscore (e.g. `_test`), which caused a master-only test to fail.

The only change from the previous commit is that I updated the regex like so:

before: `^.*TORCH_LIBRARY_IMPL_init_([^_]+)_([^_]+)_[0-9]+(\(.*)?$`
after: `^.*TORCH_LIBRARY_IMPL_init_([_]*[^_]+)_([^_]+)_[0-9]+(\(.*)?$`

I added in a `[_]*` to the beginning of the namespace capture. I did the same for the `_FRAGMENT` regex.

Verified that running `ANALYZE_TEST=1 tools/code_analyzer/build.sh` (as the master-only test does) produces no diff in the output.


Fixing regex pattern to allow for underscores at the beginning of the
namespace

This reverts commit 3c936ecd3c68f395dad01f42935f20ed8068da02.

Differential Revision: [D25123295](https://our.internmc.facebook.com/intern/diff/D25123295/)
ghstack-source-id: 117457139