[wip] add option to leave graph inputs and/or outputs quantized
Summary:

Before this PR, there was an assumption that all graph inputs
and outputs are in floating point, with some exceptions for
`standalone_module`.

This PR adds an option to specify either inputs or outputs
as being quantized.

This is useful for incremental migrations of models using Eager mode.

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 082fed9c7d12eb9a729f1b9f47d958e476848695
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48624