[PyTorch] Add VariableVersion&& overload for TensorImpl::shallow_copy_and_detach
This should reduce reference counting traffic when creating views.

The code duplication here is unfortunate and I'm open to suggestions on how to reduce it. It's especially regrettable that we create a footgun for subclasses of TensorImpl: they can accidentally override only one of the two overloads and get confusing behavior.

Differential Revision: [D25259741](https://our.internmc.facebook.com/intern/diff/D25259741/)

**NOTE FOR REVIEWERS**: This PR has internal Facebook specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D25259741/)!

ghstack-source-id: 117613147
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48681